+++ backend/CarteiraVacinacaoBE/src/main/java/com/cvoadm/CarteiraVacinacaoBE/repository/FuncionarioRepository.java
@@ -3,5 +3,5 @@ package com.cvoadm.CarteiraVacinacaoBE.repository;
 import com.cvoadm.CarteiraVacinacaoBE.model.Funcionario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Integer> {
 }
