+++ backend/CarteiraVacinacaoBE/src/main/java/com/cvoadm/CarteiraVacinacaoBE/repository/VacinaRepository.java
@@ -3,5 +3,5 @@ package com.cvoadm.CarteiraVacinacaoBE.repository;
 import com.cvoadm.CarteiraVacinacaoBE.model.Vacina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacinaRepository extends JpaRepository<Vacina, Long> {
+public interface VacinaRepository extends JpaRepository<Vacina, Integer> {
 }
